requirements.txt
torch
transformers
gradio
Pillow
matplotlib
numpy
